refact: ImageScreen Section extracts hooks, comp., cleans up
